feat(storybook): move stories to source file
