+++ packages/components/copy/xm-copy-icon.stories.ts
@@ -5,7 +5,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { ClipboardModule } from '@angular/cdk/clipboard';
 import { CommonModule } from '@angular/common';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { XmCopyIconComponent, XmCopyIconOptions } from '@xm-ngx/components/copy';
+import { XmCopyIconComponent, XmCopyIconOptions } from '@xm-ngx/components/copy/index';
 import { ControlErrorModule } from '@xm-ngx/components/control-error';
 import { XM_VALIDATOR_PROCESSING_CONTROL_ERRORS_TRANSLATES } from '@xm-ngx/components/validator-processing';
 import { XmTranslationTestingModule } from '@xm-ngx/translation/testing';
