+++ packages/components/date/xm-date-range-filter-control.stories.ts
@@ -3,7 +3,7 @@
 import { moduleMetadata } from '@storybook/angular';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
-import { DateRangeFilterControl, XmDateComponent } from '@xm-ngx/components/date';
+import { DateRangeFilterControl, XmDateComponent } from '@xm-ngx/components/date/index';
 import { XmTranslationTestingModule } from '@xm-ngx/translation/testing';
 import { ControlErrorModule } from '@xm-ngx/components/control-error';
 import { XM_VALIDATOR_PROCESSING_CONTROL_ERRORS_TRANSLATES } from '@xm-ngx/components/validator-processing';
