+++ packages/components/link/xm-link-view-copy.stories.ts
@@ -1,19 +1,21 @@
 import { CommonModule } from '@angular/common';
-import { MatIconModule } from '@angular/material/icon';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 import { Meta, moduleMetadata, Story } from '@storybook/angular';
 import { ControlErrorModule } from '@xm-ngx/components/control-error';
+import { XmCopyIconComponent } from '@xm-ngx/components/copy';
 import { HintModule } from '@xm-ngx/components/hint';
-import { XmLink } from '@xm-ngx/components/link';
+import { XmLinkViewCopyComponent } from '@xm-ngx/components/link/index';
+import { XmTextViewModule } from '@xm-ngx/components/text';
 import { XM_VALIDATOR_PROCESSING_CONTROL_ERRORS_TRANSLATES } from '@xm-ngx/components/validator-processing';
 import { XmTranslationModule } from '@xm-ngx/translation';
 import { XmTranslationTestingModule } from '@xm-ngx/translation/testing';
 
 
 export default {
-    title: 'Core/Presentation/Link/Default',
-    component: XmLink,
+    title: 'Core/Presentation/Link/View copy',
+    component: XmLinkViewCopyComponent,
     decorators: [
         moduleMetadata({
             imports: [
@@ -24,9 +26,11 @@ export default {
                 ControlErrorModule.forRoot({errorTranslates: XM_VALIDATOR_PROCESSING_CONTROL_ERRORS_TRANSLATES}),
 
                 CommonModule,
-                RouterTestingModule,
+                XmTextViewModule,
                 XmTranslationModule,
-                MatIconModule,
+                RouterTestingModule,
+                XmCopyIconComponent,
+                MatSnackBarModule,
             ],
         }),
     ],
@@ -35,8 +39,8 @@ export default {
     },
 } as Meta;
 
-const Template: Story<XmLink> = (args: XmLink) => ({
-    component: XmLink,
+const Template: Story<XmLinkViewCopyComponent> = (args: XmLinkViewCopyComponent) => ({
+    component: XmLinkViewCopyComponent,
     props: args,
 });
 
@@ -51,29 +55,22 @@ const entity = {
 export const Default = Template.bind({});
 Default.args = {
     value: entity,
-};
-
-
-export const WithTitle = Template.bind({});
-WithTitle.args = {
-    value: entity,
     config: {
-        valueTitle: 'User ID',
-    },
-};
-
-export const CustomField = Template.bind({});
-CustomField.args = {
-    value: entity,
-    config: {
+        title: 'Title',
         valueField: 'name',
     },
 };
 
-export const CustomNestedField = Template.bind({});
-CustomNestedField.args = {
+export const CustomCopy = Template.bind({});
+CustomCopy.args = {
     value: entity,
     config: {
-        valueField: 'data.country',
+        title: 'Title',
+        valueField: 'name',
+        icon: 'person',
+        copy: {
+            template: 'It is an id: {{value.id}}',
+            copiedMessage: 'Id copied to clipboard',
+        }
     },
 };
