+++ packages/components/link/xm-link-view.stories.ts
@@ -5,7 +5,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { Meta, moduleMetadata, Story } from '@storybook/angular';
 import { ControlErrorModule } from '@xm-ngx/components/control-error';
 import { HintModule } from '@xm-ngx/components/hint';
-import { XmLink, XmLinkViewComponent } from '@xm-ngx/components/link';
+import { XmLink, XmLinkViewComponent } from '@xm-ngx/components/link/index';
 import { XmTextViewModule } from '@xm-ngx/components/text';
 import { XM_VALIDATOR_PROCESSING_CONTROL_ERRORS_TRANSLATES } from '@xm-ngx/components/validator-processing';
 import { XmTranslationTestingModule } from '@xm-ngx/translation/testing';
